+++ getl/fileregistry/fileregistry_utils.py
@@ -16,8 +16,7 @@ def update_date_lifted(
 ) -> None:
     """Update date lifted column with the current date."""
     delta_table.delta_table.update(
-        F.col("date_lifted").isNull(),
-        {"date_lifted": "'{}'".format(str(datetime.now()))},
+        F.col("date_lifted").isNull(), {"date_lifted": F.lit(datetime.now())},
     )
     # Optimze and vacuum Databricks tables
     try: