+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- The `fileregistry::delta_diff` fileregistry for delta files
 
 ## [1.9.2] - 2020-12-16
 ### Added